This the first generation version of Vision
The idea was inspired from bats, who also use sound waves of high frequency to move.According to an estimate, there are 285 million visually impaired people in the world. The loss of one the most important human senses causes them a lot of hardships in daily life. The affected ones have been using the traditional white cane since many years which although being effective, still has a lot of disadvantages.So the aim of this project is to develop a cheap, yet more efficient way to help the visually impaired navigate with more speed and confidence .